Streamlit/requirements.txt
absl-py==2.3.1
altair @ file:///C:/b/abs_0b1avjyzoc/croot/altair_1743016738596/work
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beartype==0.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
captum @ file:///C:/b/abs_b8sr7c4go8/croot/captum_1717518861171/work
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click @ file:///C:/b/abs_bbg7qqz4dk/croot/click_1754477951021/work
cloudpickle @ file:///C:/b/abs_fdydcrc9_2/croot/cloudpickle_1751309451995/work
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
decorit==0.2.0
defusedxml==0.7.1
Deprecated @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/deprecated_1699499370767/work
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers @ file:///C:/b/abs_7f2_85svxz/croot/python-flatbuffers_1722369110965/work
fonttools==4.47.2
fqdn==1.5.1
fsspec==2025.7.0
gast==0.6.0
gitdb @ file:///C:/b/abs_1bkmriimss/croot/gitdb_1753362133190/work
GitPython @ file:///C:/b/abs_2bkslnqz4i/croot/gitpython_1720455044865/work
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio @ file:///C:/b/abs_4cup9zfkvm/croot/imageio_1738160462075/work
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.10.0
kiwisolver==1.4.5
lazy_loader @ file:///C:/b/abs_3fs2i5w5p3/croot/lazy_loader_1718176758844/work
lime @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/lime_1707444611909/work
llvmlite @ file:///C:/b/abs_da51wefv9a/croot/llvmlite_1741209883316/work
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.1
mpi4py @ file:///C:/b/abs_2c72fs0pz9/croot/mpi4py_1749458512296/work
mpmath==1.3.0
namex @ file:///C:/b/abs_cdq28t3827/croot/namex_1715206998407/work
narwhals @ file:///C:/b/abs_224ub5a528/croot/narwhals_1742845975749/work
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook_shim==0.2.4
numba @ file:///C:/b/abs_8cc9jdn1d8/croot/numba_1750823429225/work
numexpr @ file:///C:/b/abs_4e0h6i5ip2/croot/numexpr_1752521729801/work
numpy==1.26.4
opencv-python==4.12.0.88
opencv-python-headless==4.10.0
opentelemetry-api @ file:///C:/b/abs_f1q379y4o0/croot/opentelemetry-api_1743171676016/work
opt_einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==23.2
pandas @ file:///C:/b/abs_d5idbck212/croot/pandas_1752872294545/work/dist/pandas-2.3.1-cp312-cp312-win_amd64.whl#sha256=d5b95a45c4def734383a0863bb6421f8d6fc77e914fb01907813f55e3b1325b4
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.8
prettypyplot==0.10.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow @ file:///C:/b/abs_0aneze9pgl/croot/pyarrow_1738229440651/work/python
pycparser==2.22
Pygments==2.19.2
pyparsing==3.1.1
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.8.2
python-json-logger==3.3.0
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
scikit-image @ file:///C:/b/abs_10rg9ps1jh/croot/scikit-image_1750419495144/work
scikit-learn==1.7.1
scipy==1.16.0
seaborn @ file:///C:/b/abs_0bv5x0ucxb/croot/seaborn_1749110315600/work
setuptools==72.1.0
shap @ file:///C:/b/abs_145kiu62xt/croot/shap_1747299916224/work
six==1.16.0
sklearn-compat @ file:///C:/b/abs_2495oc29im/croot/sklearn-compat_1738003941903/work
slicer @ file:///C:/b/abs_9azh8xikg2/croot/slicer_1747291662812/work
smmap @ file:///C:/b/abs_85z04amxrn/croot/smmap_1752748960667/work
soupsieve==2.7
stack-data==0.6.3
streamlit @ file:///C:/b/abs_ebeb_1jk3u/croot/streamlit-split_1752749484326/work/streamlit
streamlit-image-select==0.6.0
sympy==1.14.0
tenacity @ file:///C:/b/abs_1749lsqys6/croot/tenacity_1730304400080/work
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_cpu==2.18.1
termcolor==3.1.0
terminado==0.18.1
tf-keras-vis==0.8.7
threadpoolctl==3.6.0
tifffile @ file:///C:/b/abs_8aikas7x_4/croot/tifffile_1741164551582/work
tinycss2==1.4.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==2.7.1
torchvision==0.22.1
tornado @ file:///C:/b/abs_7dzpc171lf/croot/tornado_1748956950306/work
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog @ file:///C:/b/abs_29_7vso69r/croot/watchdog_1740409940307/work
wcwidth @ file:///C:/b/abs_0c8p4c33bp/croot/wcwidth_1750352902378/work
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt @ file:///C:/b/abs_68gzx33lzh/croot/wrapt_1736543196433/work
yarg==0.1.9
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
